+++ SignNow.Net/Model/SignNowDocument.Fields.cs
@@ -57,11 +57,11 @@ namespace SignNow.Net.Model
             switch (fieldMeta?.Type)
             {
                 case FieldType.Text:
-                case FieldType.Dropdown:
+                case FieldType.Enumeration:
                     return Texts.FirstOrDefault(txt => txt.Id == fieldMeta.ElementId);
 
                 case FieldType.Signature:
-                case FieldType.Initial:
+                case FieldType.Initials:
                     return Signatures.FirstOrDefault(sig => sig.Id == fieldMeta.ElementId);
 
                 case FieldType.Hyperlink: